Bugfix on record access check (task #3267)
Currently we first check if user has owner type capabilities to the
current action and we modify the query accordingly.

This breaks the expected behavior for users like admins. Admins by
default they are assigned all system capabilities and since we are
first checking for owner type capabilities we modify the query to
return only records owned (assigned) to the admin user, which is
wrong.

To fix the issue explained above we just change the order of the
access check to first check for full access type capabilities, and
if the user has such access we just skip modifying the query all
together.